test(messaging): split echo fixture from messaging-handler
Commit 70f1fe5 changed the messaging-handler fixture's WIT to
import wasmcloud:messaging/consumer so the integration tests could
round-trip a reply via consumer::publish. That broke
engine::workload::tests::test_host_interface_redundancy, which
loads messaging-handler.wasm and asserts its specific import
surface (wasi:logging only) drives plugin binding — that's the
whole point of the redundancy assertion.

Reverts messaging-handler to its original wasi:logging-only form
and adds a new messaging-echo fixture for the integration tests:
imports wasmcloud:messaging/consumer, exports the handler, and
echoes the request body back on reply_to.

Wires it into both the fixtures workspace (Cargo.toml) and
build.rs's P2_FIXTURES list so the cargo build pipeline produces
tests/wasm/messaging_echo.wasm. Renames the integration tests'
constant from MESSAGING_HANDLER_WASM to MESSAGING_ECHO_WASM to
match. Uses package wasmcloud:echo / world echo in the new
fixture's WIT to avoid namespace collision with messaging-handler.
