+++ crates/wash-runtime/tests/fixtures/Cargo.toml
@@ -10,6 +10,7 @@ members = [
     "http-webgpu",
     "cpu-usage-service",
     "messaging-handler",
+    "messaging-echo",
     "inter-component-call-caller",
     "inter-component-call-callee",
     "inter-component-call-middleware",
