+++ crates/wash-runtime/tests/fixtures/messaging-echo/wit/world.wit
@@ -1,0 +1,6 @@
+package wasmcloud:echo;
+
+world echo {
+    import wasmcloud:messaging/consumer@0.2.0;
+    export wasmcloud:messaging/handler@0.2.0;
+}
