+++ crates/wash-runtime/tests/fixtures/messaging-handler/wit/world.wit
@@ -1,6 +1,6 @@
 package wasmcloud:hello;
 
 world hello {
-    import wasmcloud:messaging/consumer@0.2.0;
+    import wasi:logging/logging@0.1.0-draft;
     export wasmcloud:messaging/handler@0.2.0;
 }
